Add geoimagedelete + change image saving method + new options + refactoring/bugfixes

+++ src/commands/geostop.ts
@@ -1,16 +1,15 @@
-import { ChatInputCommandInteraction, MessageFlags, PermissionsBitField, SlashCommandBuilder } from 'discord.js';
+import { ChatInputCommandInteraction, MessageFlags, SlashCommandBuilder, TextChannel } from 'discord.js';
 import quizService from '../services/quizService';
+import { permissionCheck } from '../util/interactionHelpers';
 
 export const data = new SlashCommandBuilder()
   .setName('geostop')
   .setDescription('Stops the current quiz (if there is one)');
 
 export async function execute(interaction: ChatInputCommandInteraction) {
-  if (!(interaction.member?.permissions as PermissionsBitField).has([PermissionsBitField.Flags.KickMembers])) {
-    return await interaction.reply({
-      content: 'Permission denied',
-      flags: MessageFlags.Ephemeral,
-    });
+  const perms = await permissionCheck(interaction);
+  if (perms !== true) {
+    return;
   }
 
   try {
@@ -20,7 +19,14 @@ export async function execute(interaction: ChatInputCommandInteraction) {
       const [update] = await quizService.stopQuiz(runningQuizes[0].id);
 
       if (update.affectedRows && runningQuizes[0].message_id) {
-        const msg = await interaction.channel?.messages.fetch(runningQuizes[0].message_id);
+        const channel = (await interaction.guild?.channels.fetch(runningQuizes[0].channel_id)) as TextChannel;
+        if (!channel) {
+          return await interaction.reply({
+            content: 'Error retrieving channel',
+            flags: MessageFlags.Ephemeral,
+          });
+        }
+        const msg = await channel.messages.fetch(runningQuizes[0].message_id);
         if (msg) {
           await msg.delete();
         }
